requirements.txt
python-jenkins
dotenv